+++ files/system/usr/libexec/secureblue/utils/__init__.py
@@ -15,106 +15,20 @@
 # limitations under the License.
 
 """
-Utils for system auditing.
+Various utility functions used in secureblue scripts.
 """
 
-import asyncio
-import enum
-import os
-import re
-
-# All subprocess calls we make have trusted inputs and do not use shell=True.
 import subprocess  # nosec
 import sys
 import textwrap
 from collections.abc import Iterable
-from typing import Final
 
 import rpm
-from auditor import AuditError, Status, gettext_marker
-
-PASS: Final = Status.PASS
-INFO: Final = Status.INFO
-WARN: Final = Status.WARN
-FAIL: Final = Status.FAIL
-UNKNOWN: Final = Status.UNKNOWN
-
-
-_: Final = gettext_marker()
 
 
 def print_err(text: str) -> None:
     """Print text to stderr in bold and red."""
     print(f"\x1b[1m\x1b[31m{text}\x1b[0m", file=sys.stderr)
-
-
-def warn_if_root() -> None:
-    """If run as root, warn that this is not recommended."""
-    if not os.getuid():
-        print_err("\n" + _("*** WARNING: Running audit script as root is not recommended. ***"))
-        print_err(_("*** Some results may be misleading or incomplete. ***") + "\n")
-
-
-def get_width() -> int:
-    """Get the width in columns to be used for reports."""
-    try:
-        width = min(max(80, os.get_terminal_size().columns), 100)
-    except OSError:
-        width = 80
-    return width
-
-
-def _format_legend_entry(status: Status, description: str, width: int = 80) -> str:
-    """Format legend entry"""
-    key_str = f"[{status.to_str_in_color()}]: "
-    key_str_width = len(status.name) + 4
-    description = re.sub(r"\s+", " ", description.strip())
-    lines = textwrap.wrap(description, width=width - key_str_width)
-    if not lines:
-        return f"{key_str}\n"
-    entry = f"{key_str}{lines[0]}\n"
-    for line in lines[1:]:
-        entry += f"{' ' * key_str_width}{line}\n"
-    return entry
-
-
-def get_legend(width: int = 80) -> str:
-    """Get legend to be printed with --help option."""
-    legend = _("The following status indicators accompany checks run by the audit script:")
-    legend += "\n\n"
-    status_descriptions: dict[Status, str] = {
-        FAIL: _("check failed - the configuration may be less secure."),
-        WARN: _("partial failure, or less significant issue detected."),
-        PASS: _("check passed - no problems detected."),
-        UNKNOWN: _("unable to perform check (usually due to a file permission issue)."),
-    }
-    for status, desc in status_descriptions.items():
-        legend += _format_legend_entry(status, desc, width)
-    legend += "\n"
-    legend += _("For flatpak checks, the status indicators have more specific meanings:")
-    legend += "\n\n"
-    flatpak_status_descriptions: dict[Status, str] = {
-        FAIL: _("""app has permissions that can be used as sandbox escapes, allow it to modify
-            its own permissions, or otherwise grant very broad access to the system (e.g. access
-            to certain directories, direct D-Bus access, X11)."""),
-        WARN: _("""app has permissions that have some sandbox escape potential or otherwise
-            weaken security (e.g. PulseAudio, Bluetooth, not using hardened_malloc)."""),
-        INFO: _("""no potential sandbox escapes detected but some permissions could increase
-            attack surface or have privacy implications (e.g. network access)."""),
-        PASS: _("no app permissions flagged (however, not all permissions are audited)."),
-    }
-    for status, desc in flatpak_status_descriptions.items():
-        legend += _format_legend_entry(status, desc, width)
-    legend += "\n" + textwrap.fill(
-        textwrap.dedent(
-            _("""\
-            Note that some flatpak apps require broad permissions to function. Permissions being
-            flagged by the audit script do not necessarily mean that action should be taken.
-            """)
-        ),
-        width=width,
-    )
-    return legend
 
 
 def command_stdout(*args: str, check: bool = True) -> str:
@@ -124,30 +38,8 @@ def command_stdout(*args: str, check: bool = True) -> str:
     return subprocess.run(args, capture_output=True, check=check, text=True).stdout.strip()  # nosec
 
 
-class AsyncProcessError(AuditError):
-    """An asynchronous subprocess command returned a nonzero exit code."""
-
-
-async def async_command_stdout(cmd: str, *args: str, check: bool = True) -> str:
-    """Asynchronously run a command in the shell and return the contents of stdout."""
-    # nosemgrep: dangerous-subprocess-use-audit, dangerous-asyncio-create-exec-audit
-    sub = await asyncio.create_subprocess_exec(
-        cmd, *args, stdout=subprocess.PIPE, stderr=subprocess.DEVNULL
-    )
-    await sub.wait()
-    # pylint: disable=use-implicit-booleaness-not-comparison-to-zero
-    if check and sub.returncode != 0:
-        err = f"async command `{cmd} {' '.join(args)}` returned nonzero exit code {sub.returncode}"
-        raise AsyncProcessError(err)
-    if sub.stdout is None:
-        err = f"Failed to get stdout for async command `{cmd} {' '.join(args)}`"
-        raise AsyncProcessError(err)
-    output = await sub.stdout.read()
-    return output.decode("utf-8", errors="replace").strip()
-
-
 def command_succeeds(*args: str) -> bool:
-    """Run a command in the shell and return the contents of stdout."""
+    """Run a command in the shell and return whether it completes with return code 0."""
     # We only call this with trusted inputs and do not set shell=True.
     # nosemgrep: dangerous-subprocess-use-audit
     return subprocess.run(args, capture_output=True, check=False).returncode == 0  # nosec
@@ -177,63 +69,13 @@ def is_rpm_package_installed(name: str) -> bool:
     return len(matches) > 0
 
 
-class Image(enum.Enum):
-    """Fedora atomic base image"""
-
-    SILVERBLUE = enum.auto()
-    KINOITE = enum.auto()
-    SERICEA = enum.auto()
-    COSMIC = enum.auto()
-    COREOS = enum.auto()
-    IOT = enum.auto()
-
-    @classmethod
-    def from_image_ref(cls, image_ref: str) -> "Image | None":
-        """Convert an image reference to the corresponding Image enum instance."""
-        image_dict = {
-            "silverblue": cls.SILVERBLUE,
-            "kinoite": cls.KINOITE,
-            "sericea": cls.SERICEA,
-            "cosmic": cls.COSMIC,
-            "securecore": cls.COREOS,
-            "iot": cls.IOT,
-        }
-        image_name = image_ref.rsplit("/", maxsplit=1)[-1]
-        image_prefix = image_name.split("-", maxsplit=1)[0]
-        return image_dict.get(image_prefix)
-
-    def is_server(self) -> bool:
-        """Is the image a server image?"""
-        return self in (Image.COREOS, Image.IOT)
-
-    def is_desktop(self) -> bool:
-        """Is the image a desktop image?"""
-        return not self.is_server()
-
-
-async def get_flatpak_permissions(name: str, version: str) -> str:
-    """Get permissions for an installed flatpak."""
-    return await async_command_stdout("flatpak", "info", "--show-permissions", name, version)
-
-
-def validate_sysctl(sysctl: str, actual: str, expected: str) -> bool:
-    """Validate a sysctl value against an expected value."""
-    actual = re.sub(r"\s+", " ", actual.strip())
-    replace = {"disabled": "0", "enabled": "1"}.get(actual)
-    if replace is not None:
-        actual = replace
-    if sysctl == "kernel.sysrq":
-        # Both 0 and 4 are secure values for this setting. For details, see:
-        # https://www.kernel.org/doc/html/latest/admin-guide/sysrq.html
-        return actual in (expected, "0", "4")
-    return actual == expected
-
-
 def is_using_vpn() -> bool:
     """Returns whether an OpenVPN or Wireguard VPN is currently in use."""
 
     # Check for Wireguard VPN use.
     wg_out = command_stdout("/usr/bin/ip", "link", "show", "type", "wireguard")
+    if wg_out:
+        return True
 
     # For OpenVPN, we need to figure out whether the default route is via a TUN/TAP interface.
     # Otherwise, we'd detect virtual networks, etc.
@@ -247,7 +89,7 @@ def is_using_vpn() -> bool:
             has_openvpn = True
             break
 
-    return wg_out or has_openvpn
+    return has_openvpn
 
 
 def interruptible_ask(prompt: str) -> str:
